seperates .hpp files in include directory into .cpp files
